доработка, переделываю контроллер по ресту убираю map из RequestBody, заменяю на сущности

+++ src/main/java/com/tradeservice/testproject/services/impl/GoodsServiceImpl.java
@@ -4,7 +4,6 @@ import com.tradeservice.testproject.entities.Goods;
 import com.tradeservice.testproject.repositories.GoodsRepository;
 import com.tradeservice.testproject.services.GoodsService;
 import java.util.List;
-import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,6 +36,9 @@ public class GoodsServiceImpl implements GoodsService {
 
   @Override
   public void delete(Long id) {
+    if (id == null) {
+      throw new NullPointerException("Не указан goodsId");
+    }
     goodsRepository.deleteById(id);
   }
 
@@ -47,6 +49,9 @@ public class GoodsServiceImpl implements GoodsService {
 
   @Override
   public Goods edit(Goods goods) {
+    if (goods.getGoodsId() == null) {
+      throw new NullPointerException("Не указан goodsId");
+    }
     return goodsRepository.saveAndFlush(goods);
   }
 
@@ -54,28 +59,4 @@ public class GoodsServiceImpl implements GoodsService {
   public List<Goods> getAll() {
     return goodsRepository.findAll();
   }
-
-
-  // Чтобы изменить можно было по имени. нужно сначала найти товар в БД по имени, затем изменить
-  public Goods editWithoutId(Map<String, String> map) {
-    String oldName = map.get("oldName");
-    if (oldName == null) {
-      throw new IllegalArgumentException("Необходим параметр oldName! ");
-    }
-    String newName = map.get("newName");
-    String strNewPrice = map.get("newPrice");
-
-    Goods result = getByName(oldName);
-    if (result == null) {
-      throw new NoSuchElementException("Такого товара нет! ");
-    }
-      if (newName != null) {
-        result.setName(newName);
-      }
-      if (strNewPrice != null) {
-        Double newPrice = Double.parseDouble(strNewPrice);
-        result.setPrice(newPrice);
-      }
-      return goodsRepository.saveAndFlush(result);
-    }
 }
